requirements.txt
streamlit
psycopg2
sqlalchemy
python-dotenv
pandas
matplotlib
langchain-groq

-e .